feat(core/embed): prepare usb driver for low power mode
[no changelog]
