+++ core/site_scons/models/T3T1/trezor_t3t1_revE.py
@@ -100,14 +100,14 @@ def configure(
 
     if "usb" in features_wanted:
         sources += [
-            "embed/trezorhal/stm32u5/usb/usbd_class_hid.c",
-            "embed/trezorhal/stm32u5/usb/usbd_class_vcp.c",
-            "embed/trezorhal/stm32u5/usb/usbd_class_webusb.c",
+            "embed/trezorhal/stm32u5/usb/usb_class_hid.c",
+            "embed/trezorhal/stm32u5/usb/usb_class_vcp.c",
+            "embed/trezorhal/stm32u5/usb/usb_class_webusb.c",
+            "embed/trezorhal/stm32u5/usb/usb.c",
             "embed/trezorhal/stm32u5/usb/usbd_conf.c",
             "embed/trezorhal/stm32u5/usb/usbd_core.c",
             "embed/trezorhal/stm32u5/usb/usbd_ctlreq.c",
             "embed/trezorhal/stm32u5/usb/usbd_ioreq.c",
-            "embed/trezorhal/stm32u5/usb/usbd.c",
             "vendor/stm32u5xx_hal_driver/Src/stm32u5xx_ll_usb.c",
         ]
         features_available.append("usb")
